devided main.js into main and route
